+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
 * Added support for registering a custom timing by providing a number describing the elapsed number of milliseconds.
 * Improved performance when using OkanshiMonitor to get monitors
 * Ensured thread safety in metric types by using locks where needed
+* Added new gauge, AverageGauge, for calculating average value for a specific interval
 
 ### 4.0.0-beta